Add parsing of Phred scores from FastQ file
Added function to retrieve the Phred scores from FastQ reads, so that
their values may be averaged.

The function takes every 4th line (containing the Phred scores) from a
FastQ file, using GNU sed to stream the file, and python subprocess to
pipe sed's output into a list.

This code is a bit more convoluted than a simple counter loop, but it
shaves off a couple of seconds of execution time.