Add preallocation of Phred score NumPy matrix
To save time one execution, a NumPy matrix is initialized with
with NaN values, using dimensions: (number of reads, longest read length).

This way the bases for every read have an adress in the matrix, no
append operations wil have to be performed, and vector arithmatic can
rapidly be performed to calculate averages.

Adds dependency for NumPy